Finished services, started console app

+++ CreateTables.sql
@@ -9,7 +9,7 @@ CREATE TABLE Accounts(
 	AccountNumber INT NOT NULL,
 	Pin NVARCHAR(4) NOT NULL,
 	AccountType NVARCHAR(10) NOT NULL,
-	Balance BIGINT NOT NULL,
+	Balance MONEY NOT NULL,
 	CustomerID INT NOT NULL,
 	CONSTRAINT [FK_dbo.Accounts_dbo.Customers_CustomerID] FOREIGN KEY ([CustomerID]) REFERENCES [dbo].[Customers] ([CustomerID]) ON DELETE CASCADE
 );
@@ -22,14 +22,14 @@ CREATE TABLE Transactions(
 
 CREATE TABLE Withdrawls(
 	WithdrawlID INT IDENTITY PRIMARY KEY NOT NULL,
-	Amount BIGINT NOT NULL,
+	Amount MONEY NOT NULL,
 	TransactionID INT NOT NULL,
 	CONSTRAINT [FK_dbo.Withdrawls_dbo.Transactions_TransactionID] FOREIGN KEY ([TransactionID]) REFERENCES [dbo].[Transactions] ([TransactionID]) ON DELETE CASCADE,
 );
 
 CREATE	TABLE Deposits(
 	DepositID INT IDENTITY PRIMARY KEY NOT NULL,
-	Amount BIGINT NOT NULL,
+	Amount MONEY NOT NULL,
 	TransactionID INT NOT NULL,
 	CONSTRAINT [FK_dbo.Deposits_dbo.Transactions_TransactionID] FOREIGN KEY ([TransactionID]) REFERENCES [dbo].[Transactions] ([TransactionID]) ON DELETE CASCADE,
 );
